chore: add MIT LICENSE + SPDX headers across examples
Adds a top-level MIT LICENSE (copyright 2026 Chris Collins) and prepends
SPDX-License-Identifier + copyright headers to every source, config,
build, and docs file across the example projects — 328 files touched
by the header-injection script from the main repo.

Extends .gitignore to exclude the per-example binary targets that were
missing (ballblazer, fractalus, ham_raster, ham_test, masterwork, opus,
organizer, terrain_test, void_trader) plus macOS *.dSYM/ debug dirs
that were leaking in from local test runs.

fractalus/ is intentionally left out of this commit — its files landed
as part of a functional dual-target port that ships as a separate
commit for reviewability.

+++ rock_blaster/game.c
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2026 Chris Collins
+
 /*
  * Rock Blaster - Game logic
  */
@@ -89,7 +92,10 @@ static ULONG rng(void)
 static Fixed rng_range(Fixed lo, Fixed hi)
 {
     ULONG r = rng();
-    return lo + (Fixed)((r * (ULONG)(hi - lo)) >> 15);
+    /* Widen to 64-bit before multiply so r * (hi - lo) can't overflow the
+     * 32-bit ULONG intermediate. r is 15-bit and (hi - lo) may be full
+     * fixed-point range. */
+    return lo + (Fixed)(((unsigned long long)r * (unsigned long long)(ULONG)(hi - lo)) >> 15);
 }
 
 /* Wrap coordinates to screen */
